third_party/WebKit/LayoutTests/fast/canvas/color-space/canvas-getImageData-linear-rgb.html

<!DOCTYPE html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(drawSRGBThenGetImageDataLinearRGB, 'tests drawing SRGB color on LinearRGB \
canvas and getting LinearRGB image data.');
test(putImageDataLinearRGBThenGetImageDataLinearRGB, 'tests putting and \
getting LinearRGB image data on LinearRGB canvas.');
test(putImageDataSRGBThenGetImageDataLinearRGB, 'tests putting SRGB image data \
on LinearRGB canvas and getting LinearRGB image data.');

function assert_array_approx_equals(actual, expected, epsilon)
{
  assert_true(actual.length === expected.length);
  for (var i=0; i < actual.length; i++)
    assert_approx_equals(actual[i], expected[i], epsilon);
}

function drawSRGBThenGetImageDataLinearRGB() {
    var canvas = document.createElement('canvas');
    canvas.width = 10;
    canvas.height = 10;
    var ctx = canvas.getContext('2d',
        {colorSpace: 'linear-rgb', pixelFormat:'float16'});
    ctx.fillStyle = 'rgb(50, 100, 150)';
    ctx.fillRect(0, 0, 10, 10);
    var pixel = ctx.getImageData(5, 5, 1, 1).dataUnion;
    // Check against the same color in LinearRGB. 0.01 protects the test against
    // color conversion deviations.
    assert_array_approx_equals(pixel, [0.03, 0.13, 0.30, 1], 0.01);
}

function putImageDataLinearRGBThenGetImageDataLinearRGB() {
    var canvas = document.createElement('canvas');
    canvas.width = 10;
    canvas.height = 10;
    var ctx = canvas.getContext('2d',
        {colorSpace: 'linear-rgb', pixelFormat:'float16'})
    var dataLinearRGB = new Float32Array(4);
    dataLinearRGB[0] = 0.03;
    dataLinearRGB[1] = 0.13;
    dataLinearRGB[2] = 0.30;
    dataLinearRGB[3] = 1;
    var imageData = ctx.createImageData(dataLinearRGB, 1, 1,
        {colorSpace: 'linear-rgb', storageFormat:'float32'});
    ctx.putImageData(imageData, 5, 5);
    var pixel = ctx.getImageData(5, 5, 1, 1).dataUnion;
    // Check against the same color in LinearRGB. 0.001 protects the test against
    // rounding errors.
    assert_array_approx_equals(pixel, [0.03, 0.13, 0.30, 1], 0.001);
}

function putImageDataSRGBThenGetImageDataLinearRGB() {
    var canvas = document.createElement('canvas');
    canvas.width = 10;
    canvas.height = 10;
    var ctx = canvas.getContext('2d',
        {colorSpace: 'linear-rgb', pixelFormat:'float16'});
    var dataSRGB = new Uint8ClampedArray(4);
    dataSRGB[0] = 50;
    dataSRGB[1] = 100;
    dataSRGB[2] = 150;
    dataSRGB[3] = 255;
    var imageData = ctx.createImageData(dataSRGB, 1, 1,
        {colorSpace: 'srgb', storageFormat:'uint8'});
    ctx.putImageData(imageData, 5, 5);
    var pixel = ctx.getImageData(5, 5, 1, 1).dataUnion;
    // Check against the same color in LinearRGB. 0.01 protects the test against
    // color conversion deviations.
    assert_array_approx_equals(pixel, [0.03, 0.13, 0.30, 1], 0.01);
}
</script>
</body>
